clean up geometries
