+++ src/lib/components/Entities/GeometryModel.svelte
