+++ src/lib/components/Entities/Mesh.svelte
@@ -1,26 +1,29 @@
 <script lang="ts">
-	import type { Pose } from '@viamrobotics/sdk'
 	import type { Entity } from 'koota'
 
 	import { T, type Props as ThrelteProps, useThrelte } from '@threlte/core'
 	import { type Snippet } from 'svelte'
-	import { Color, DoubleSide, FrontSide, Mesh, MeshToonMaterial } from 'three'
+	import { Color, DoubleSide, FrontSide, Group, Mesh, MeshToonMaterial } from 'three'
 
 	import { asColor } from '$lib/buffer'
 	import { colors, darkenColor } from '$lib/color'
 	import { traits, useTrait } from '$lib/ecs'
 	import { poseToObject3d } from '$lib/transform'
 
+	import { useEntityEvents } from './hooks/useEntityEvents.svelte'
+
 	interface Props extends Omit<ThrelteProps<Mesh>, 'ref'> {
 		entity: Entity
-		color?: string
-		center?: Pose
 		children?: Snippet
 	}
 
-	const { entity, color: overrideColor, center, children, ...rest }: Props = $props()
+	const { entity, children }: Props = $props()
 
 	const { invalidate } = useThrelte()
+
+	const worldMatrix = useTrait(() => entity, traits.WorldMatrix)
+	const center = useTrait(() => entity, traits.Center)
+	const invisible = useTrait(() => entity, traits.InheritedInvisible)
 	const name = useTrait(() => entity, traits.Name)
 	const entityColors = useTrait(() => entity, traits.Colors)
 	const entityColor = useTrait(() => entity, traits.Color)
@@ -30,10 +33,6 @@
 	const renderOrder = useTrait(() => entity, traits.RenderOrder)
 
 	const color = $derived.by(() => {
-		if (overrideColor) {
-			return overrideColor
-		}
-
 		if (entityColors.current) {
 			return asColor(entityColors.current, new Color())
 		}
@@ -48,6 +47,28 @@
 	const hasVertexColors = $derived(bufferGeometry.current?.getAttribute('color') !== undefined)
 
 	const currentOpacity = $derived(opacity.current ?? 0.7)
+
+	const events = useEntityEvents(() => entity)
+
+	const group = new Group()
+	group.matrixAutoUpdate = false
+
+	$effect(() => {
+		if (!worldMatrix.current) return
+
+		group.matrix.copy(worldMatrix.current)
+
+		/**
+		 * Keep position/quaternion/scale in sync with matrix so TransformControls
+		 * (which reads/writes those fields) sees the entity's actual transform on
+		 * drag start. Without this, the gizmo applies its drag delta against an
+		 * identity baseline and the frame snaps to identity on first onChange.
+		 */
+		group.matrix.decompose(group.position, group.quaternion, group.scale)
+
+		group.updateMatrixWorld()
+		invalidate()
+	})
 
 	const material = new MeshToonMaterial()
 
@@ -65,47 +86,52 @@
 	const mesh = new Mesh()
 
 	$effect(() => {
-		if (center) {
-			poseToObject3d(center, mesh)
+		if (center.current) {
+			poseToObject3d(center.current, mesh)
 			invalidate()
 		}
 	})
 </script>
 
 <T
-	is={mesh}
-	name={entity}
-	userData.name={name}
-	renderOrder={renderOrder.current}
-	{...rest}
+	is={group}
+	visible={invisible.current !== true}
 >
-	{#if bufferGeometry.current}
-		<T is={bufferGeometry.current}>
-			{#snippet children({ ref: geo })}
-				<!--
+	<T
+		is={mesh}
+		name={entity}
+		userData.name={name}
+		renderOrder={renderOrder.current}
+		{...events}
+	>
+		{#if bufferGeometry.current}
+			<T is={bufferGeometry.current}>
+				{#snippet children({ ref: geo })}
+					<!--
 					TODO(mp) currently some bufferGeometries are coming in empty,
 					this is a quick fix but this should be handled upstream
 				-->
-				{#if geo.getAttribute('position').array.length > 0}
-					<T.LineSegments
-						raycast={() => null}
-						bvh={{ enabled: false }}
-					>
-						<T.EdgesGeometry args={[geo, 0]} />
-						<T.LineBasicMaterial color={darkenColor(color, 10)} />
-					</T.LineSegments>
-				{/if}
-			{/snippet}
-		</T>
-	{/if}
+					{#if geo.getAttribute('position').array.length > 0}
+						<T.LineSegments
+							raycast={() => null}
+							bvh={{ enabled: false }}
+						>
+							<T.EdgesGeometry args={[geo, 0]} />
+							<T.LineBasicMaterial color={darkenColor(color, 10)} />
+						</T.LineSegments>
+					{/if}
+				{/snippet}
+			</T>
+		{/if}
 
-	<T
-		is={material}
-		color={hasVertexColors ? 0xffffff : color}
-		vertexColors={hasVertexColors}
-		side={bufferGeometry.current ? DoubleSide : FrontSide}
-		depthTest={materialProps.current?.depthTest ?? true}
-	/>
+		<T
+			is={material}
+			color={hasVertexColors ? 0xffffff : color}
+			vertexColors={hasVertexColors}
+			side={bufferGeometry.current ? DoubleSide : FrontSide}
+			depthTest={materialProps.current?.depthTest ?? true}
+		/>
 
-	{@render children?.()}
+		{@render children?.()}
+	</T>
 </T>
